+++ microservice-core/include/settings/IServerSettings.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <cstddef>
 
 /**
  * @file IServerSettings.hpp
@@ -12,4 +13,5 @@ public:
     virtual ~IServerSettings() = default;
     virtual std::string getHost() const = 0;
     virtual int getPort() const = 0;
+    virtual size_t getMaxRequestBodySize() const = 0;
 };